Добавляет фотографии других пользователей
